+++ backend/libraries/stable_memory_map/src/keys.rs
@@ -15,6 +15,7 @@ mod message_id;
 mod principal;
 mod search_index;
 mod storage;
+mod thread_read;
 mod token_swap;
 mod user_id;
 mod user_metrics;
@@ -30,6 +31,7 @@ pub use message_id::*;
 pub use principal::*;
 pub use search_index::*;
 pub use storage::*;
+pub use thread_read::*;
 pub use token_swap::*;
 pub use user_id::*;
 pub use user_metrics::*;
@@ -153,7 +155,8 @@ pub enum KeyType {
     MessageActivityEvent = 47,
     MessageActivityEventId = 48,
     ChitEvent = 49,
-    // 50 and 51 are reserved for threads read
+    GroupThreadRead = 50,
+    ChannelThreadRead = 51,
     TokenSwap = 52,
     #[cfg(test)]
     TestSmallEntries = 255,
@@ -228,7 +231,9 @@ impl KeyType {
             | KeyType::ChannelSearchSender
             | KeyType::MessageActivityEvent
             | KeyType::MessageActivityEventId
-            | KeyType::ChitEvent => MapClass::SmallEntries,
+            | KeyType::ChitEvent
+            | KeyType::GroupThreadRead
+            | KeyType::ChannelThreadRead => MapClass::SmallEntries,
             #[cfg(test)]
             KeyType::TestSmallEntries => MapClass::SmallEntries,
         }
@@ -308,6 +313,8 @@ impl TryFrom<u8> for KeyType {
             47 => Ok(KeyType::MessageActivityEvent),
             48 => Ok(KeyType::MessageActivityEventId),
             49 => Ok(KeyType::ChitEvent),
+            50 => Ok(KeyType::GroupThreadRead),
+            51 => Ok(KeyType::ChannelThreadRead),
             52 => Ok(KeyType::TokenSwap),
             #[cfg(test)]
             255 => Ok(KeyType::TestSmallEntries),